Added funktion to all actionbars, Added EnergySystem

+++ spel_modul2/GameEngine/Components/CoreComponents/EnergyComponent.cs
@@ -2,10 +2,10 @@
 {
     public class EnergyComponent : IComponent
     {
-        public int Max { get; set; }
-        public int Current { get; set; }
+        public float Max { get; set; }
+        public float Current { get; set; }
 
-        public EnergyComponent(int max)
+        public EnergyComponent(float max)
         {
             Max = max;
             Current = Max;
